Move mock_time to test-helpers crate
